update Mountain entity property simaksiPrice from Double to Long

+++ src/main/java/com/abdav/giri_guide/entity/Mountains.java
@@ -62,7 +62,7 @@ public class Mountains extends AuditEntity {
     private boolean useSimaksi = false;
 
     @Builder.Default
-    private Double priceSimaksi = 0.0;
+    private Long priceSimaksi = 0L;
 
     @Nullable
     @Column(columnDefinition = "Text")
